test: resolve supabase.ts conflict

+++ src/lib/server/supabase.ts
@@ -1,37 +1,46 @@
 import { createClient } from '@supabase/supabase-js';
 // import { env } from '$env/dynamic/public';
-import { Student, type StudentResponse } from '$lib/classes/Student';
+import { Student, type StudentFilter, type StudentResponse } from '$lib/classes/Student';
 
 // creates the connection to SUSe supabase
-export const supabase = createClient("https://yfhwfzwacdlqmyunladz.supabase.co", "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6InlmaHdmendhY2RscW15dW5sYWR6Iiwicm9sZSI6ImFub24iLCJpYXQiOjE3MDk5MDIyNjEsImV4cCI6MjAyNTQ3ODI2MX0.gzr5edDIVJXS1YYsQSyuZhc3oHGQYuVDtVfH4_2d30A");
+export const supabase = createClient(
+	'https://yfhwfzwacdlqmyunladz.supabase.co',
+	'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6InlmaHdmendhY2RscW15dW5sYWR6Iiwicm9sZSI6ImFub24iLCJpYXQiOjE3MDk5MDIyNjEsImV4cCI6MjAyNTQ3ODI2MX0.gzr5edDIVJXS1YYsQSyuZhc3oHGQYuVDtVfH4_2d30A'
+);
 
-export async function selectStudentDB(
-	studentNumber: number = 0,
-	username: string = ''
-): Promise<StudentResponse> {
-	/* Selects the student record/s from the database given their student number and username. */
-	let orFilter: string = ``;
+const success = {
+	success: true,
+	studentRaws: null,
+	error: null
+};
 
-	const selectQuery = supabase.from('student').select('*');
+export async function selectStudentDB(filter: StudentFilter): Promise<StudentResponse> {
+	/* Selects the student record/s from the database using a filter.
+    Filter only contains option for student number and username for now. */
+	let minSN: number = filter.minStudentNumber;
+	let maxSN: number = filter.maxStudentNumber;
 
-	if (studentNumber) {
-		// if there is a given studentNumber, add this predicate to orFilter
-		orFilter += `sn_id.eq.${studentNumber}`;
-	}
-	if (username) {
-		if (orFilter.length) {
-			orFilter += `, `;
-		}
-		// if there is a given username, add this predicate to orFilter
-		orFilter += `username.eq.${username}`;
-	}
-
-	if (orFilter) {
-		// filters the student table for records that pass at least one of the conditions
-		selectQuery.or(orFilter);
+	if (Math.floor(minSN / 2000) == 1 && Math.floor(maxSN / 2000) == 1) {
+		// inputs are years and not specific student numbers
+		minSN = filter.minStudentNumber * 100000;
+		maxSN = (filter.maxStudentNumber + 1) * 100000;
 	}
 
-	const { data, error } = await selectQuery;
+	if ((Math.floor(minSN / 200000000) != 1 && Math.floor(maxSN / 200000000) != 1) || minSN > maxSN) {
+		// student numbers range have to be valid
+		return {
+			success: false,
+			studentRaws: null,
+			error: 'Error: Student number range invalid'
+		};
+	}
+
+	const { data, error } = await supabase
+		.from('student')
+		.select('*')
+		.gte('sn_id', minSN) // student number should be between an inclusive range
+		.lte('sn_id', maxSN)
+		.like('username', '%' + filter.username + '%'); // username can be found in any position
 
 	if (error) {
 		return {
@@ -49,7 +58,7 @@ export async function selectStudentDB(
 }
 
 export async function insertStudentDB(student: Student): Promise<StudentResponse> {
-	/* Inserts a student record into the database. */
+	/* Inserts a non-existing student record into the database. */
 	const { error } = await supabase.from('student').insert(student.toStudentDBObj());
 
 	if (error) {
@@ -60,16 +69,36 @@ export async function insertStudentDB(student: Student): Promise<StudentResponse
 		};
 	}
 
+	return success;
+}
+
+async function checkStudentExistsDB(filter: StudentFilter): Promise<StudentResponse> {
+	/* Checks if there is a single existing record of a student with the given student number and username. */
+	const studentDB = await selectStudentDB(filter);
+
+	if (studentDB.success && studentDB.studentRaws?.length == 1) {
+		return success;
+	}
+
 	return {
-		success: true,
+		success: false,
 		studentRaws: null,
-		error: null
+		error: 'Error: Student does not exist'
 	};
 }
 
 export async function updateStudentDB(student: Student): Promise<StudentResponse> {
 	/* Updates a student record based using their student number and username.
     NOTE: Cannot update the student number or username of a student. Need to delete and register again. */
+	const studentCheck = await checkStudentExistsDB({
+		minStudentNumber: student.studentNumber,
+		maxStudentNumber: student.studentNumber,
+		username: student.username
+	});
+
+	if (!studentCheck.success) {
+		return studentCheck;
+	}
 
 	const { error } = await supabase
 		.from('student')
@@ -81,7 +110,8 @@ export async function updateStudentDB(student: Student): Promise<StudentResponse
 			program: student.program,
 			phone_number: student.phoneNumber
 		})
-		.match({ sn_id: student.studentNumber, username: student.username });
+		.eq('sn_id', student.studentNumber)
+		.eq('username', student.username);
 
 	if (error) {
 		return {
@@ -91,20 +121,26 @@ export async function updateStudentDB(student: Student): Promise<StudentResponse
 		};
 	}
 
-	return {
-		success: true,
-		studentRaws: null,
-		error: null
-	};
+	return success;
 }
 
 export async function deleteStudentDB(student: Student): Promise<StudentResponse> {
-	/* Updates a student record based using their student number and username.
-    NOTE: Cannot update the student number or username of a student. Need to delete and register again. */
+	/* Deletes an existing student record. */
+	const studentCheck = await checkStudentExistsDB({
+		minStudentNumber: student.studentNumber,
+		maxStudentNumber: student.studentNumber,
+		username: student.username
+	});
+
+	if (!studentCheck.success) {
+		return studentCheck;
+	}
+
 	const { error } = await supabase
 		.from('student')
-        .delete()
-		.match({ sn_id: student.studentNumber, username: student.username });
+		.delete()
+		.eq('sn_id', student.studentNumber)
+		.eq('username', student.username);
 
 	if (error) {
 		return {
@@ -114,20 +150,26 @@ export async function deleteStudentDB(student: Student): Promise<StudentResponse
 		};
 	}
 
-	return {
-		success: true,
-		studentRaws: null,
-		error: null
-	};
+	return success;
 }
 
 export async function approveStudentDB(student: Student): Promise<StudentResponse> {
-	/* Updates a student record based using their student number and username.
-    NOTE: Cannot update the student number or username of a student. Need to delete and register again. */
+	/* Approves an existing student by updating their isEnrolled attribute to true. */
+	const studentCheck = await checkStudentExistsDB({
+		minStudentNumber: student.studentNumber,
+		maxStudentNumber: student.studentNumber,
+		username: student.username
+	});
+
+	if (!studentCheck.success) {
+		return studentCheck;
+	}
+
 	const { error } = await supabase
 		.from('student')
-        .update({ is_enrolled: true})
-		.match({ student_number: student.studentNumber, username: student.username });
+		.update({ is_enrolled: true })
+		.eq('sn_id', student.studentNumber)
+		.eq('username', student.username);
 
 	if (error) {
 		return {
@@ -137,9 +179,5 @@ export async function approveStudentDB(student: Student): Promise<StudentRespons
 		};
 	}
 
-	return {
-		success: true,
-		studentRaws: null,
-		error: null
-	};
+	return success;
 }
